update email service
