+++ app.js
@@ -18,8 +18,8 @@ app.use(logger('dev'));
 app.use(bodyParser.json({limit: '50mb'}));
 app.use(bodyParser.urlencoded({limit: '50mb', extended: true}));
 app.use(session({
-  secret: 'aptgamestempsecret123',
-  cookie: {expires: new Date(253402300000000)},  // cookie never expire
+  secret: process.env.SECRET,
+  cookie: {expires: new Date(Date.now() + 86400000)}, // expire in 1 day
   resave: true,
   saveUninitialized: true
 }));
